新增车辆的增删改查

+++ src/main/java/com/caidao/controller/system/SysMenuController.java
@@ -79,7 +79,7 @@ public class SysMenuController {
 	 * @return
 	 */
 	@GetMapping("info/{id}")
-	public ResponseEntity<SysMenu> getOneMenu(@PathVariable("id") Long id){
+	public ResponseEntity<SysMenu> getOneMenu(@PathVariable("id") Integer id){
 		SysMenu sysMenu = sysMenuService.getById(id);
 		return ResponseEntity.ok(sysMenu);
 	}
